Rename controller tests & delete redundant scenarios
Test cases were renamed to ensure consistency, as the method names were not following the accepted given_when_then structure.

Test scenarios that checked for unauthorised responses did not work as intended because the springSecurityFilterChain is disabled in controller test classes. Instead, these scenarios will be moved to the integration test classes.