Implement manage_income_evidence stored procedure logic in Evidence Service
This commit implements and refactors the logic from _manage_income_evidence_
stored procedure from the legacy database into the Evidence Service. Some of
the logic in the stored procedure has not been implemented as it is no longer
applicable or necessary, or is defined as part of a separate story, such as
migrating the create_default_income_evidence stored procedure.